Add option to download all test data concurrently
